Create a `radix-engine-toolkit/types` module.
The old `models` module was barely getting used despite the toolkit
having a lot of types that it uses. Additionally, the types were spread
across different modules such as `sbor` and `models`. This commit
moves all of the types from these two modules into a new module called
`types` which is where the majority of the toolkit types will live from
now on.