Move away from `utils` and into `extensions`
This commit is the start of a move of certain utility functions and
methods from the `utils` module and into a new module called the
`extensions` module.

Extensions allow us to define more functions and methods on types almost
as if the types are defined in our crate, they do this by using new
traits.

Eventually, in this PR, there should no longer be a utils module and
there should only be an extensions module.