Add grouped instructions.
This commit adds a new `GroupedInstruction` type which is defined in the
`types::grouped_instruction` module alongside the macro that defines it.

Everything about this newly added type is defined through a macro which
allows us to not write or maintain the 1600 lines of code that get
generated by the macro.