Add a grouped "entity type" type.
In a similar way to instructions, grouping entity types can be useful in
transaction type classification and therefore we introduce a new concept
of grouped entity types.

This concept should allow for a simple way to do checks such as "is the
entity an account" instead of needing to check against multiple
different entity types.