Add logic for traversing manifests with visitor.
This commit adds the `traverse` function to the `traverser` module of
the `manifest_analysis` module. This function contains the logic used
for traversal of manifests with a manifest analysis visitor.

This function does not accept multiple visitors. Therefore, the visitor
composition technique introduced a few commits back must be used.