Refactor the Invocations IO module.
This commit refactors the invocations IO module from what we've done in
a previous commit. An issue with the previous implementation is that the
invocation IOs were only defined if dynamic analysis was available. If
not, then we didn't compute the dynamic analysis.

This commit updates the implementation so that the invocation IOs are
computed with either just static analysis or with static and dynamic
analysis.