Remove old transaction types code
This is a pretty big commit as it does many different things that all
needed to be done together. The major thing that is done in this commit
is the removal of the old code that was used in transaction types
classification and detection. With its removal, a few things were made
possible and were changed as part of this commit too:

* The old code relied on the `schema_visitor` module and logic which we
no longer require. So, this module was removed.
* With the removal of the `schema_visitor` side of the code we no longer
need the `statics` module and therefore that was removed as well. It's
planned that this will be replaced by code that produce similar outcomes
without relying on the same set of information.
* Removed the `utils` module of the toolkit as it was only needed by the
old code and it will be replaced with the upcoming refactors.
* Removed the need for the nightly compiler in the toolkit. This change
has been made to CI, internal scripts, formatting rules, and pretty much
everywhere else where it needed to be made.
* Due to the removal of the Nightly dependency we needed to re-format
the entire codebase according to the rules of the current compiler.
Therefore, this commit introduced a lot of noise sadly.
* Re-organized some tests into various modules and disabled some tests
that we expect will fail due to some parts of the code currently being
disabled.